Refactored LDO methods into the solid library
