+++ packages/solid/src/SolidLdoDataset.ts
@@ -1,5 +1,6 @@
-import { LdoDataset } from "@ldo/ldo";
-import type { DatasetChanges, GraphNode } from "@ldo/rdf-utils";
+import type { LdoBase, ShapeType } from "@ldo/ldo";
+import { LdoDataset, startTransaction } from "@ldo/ldo";
+import type { DatasetChanges, GraphNode, SubjectNode } from "@ldo/rdf-utils";
 import type { Dataset, DatasetFactory, Quad } from "@rdfjs/types";
 import type {
   UpdateResult,
@@ -17,6 +18,7 @@ import type { SolidLdoDatasetContext } from "./SolidLdoDatasetContext";
 import { splitChangesByGraph } from "./util/splitChangesByGraph";
 import type { ContainerUri, LeafUri } from "./util/uriTypes";
 import { isContainerUri } from "./util/uriTypes";
+import type { Resource } from "./resource/Resource";
 
 export class SolidLdoDataset extends LdoDataset {
   public context: SolidLdoDatasetContext;
@@ -95,5 +97,26 @@ export class SolidLdoDataset extends LdoDataset {
             result.type === "updateSuccess",
         ),
     };
+  }
+
+  /**
+   * Shorthand for solidLdoDataset
+   *   .usingType(shapeType)
+   *   .write(...resources.map((r) => r.uri))
+   *   .fromSubject(subject);
+   * @param shapeType The shapetype to represent the data
+   * @param subject A subject URI
+   * @param resources The resources changes to should written to
+   */
+  createData<Type extends LdoBase>(
+    shapeType: ShapeType<Type>,
+    subject: string | SubjectNode,
+    ...resources: Resource[]
+  ): Type {
+    const linkedDataObject = this.usingType(shapeType)
+      .write(...resources.map((r) => r.uri))
+      .fromSubject(subject);
+    startTransaction(linkedDataObject);
+    return linkedDataObject;
   }
 }
